moved ApiError component to components folder
